Refac: MainNavigation refac
